api: all routes call python excel recalculation with the same helper function
just pass the input data and name of the python script to the helper
function. The function executes the python script synchronously (wooo no
concurrency bugs [I think]) and returns a buffer containing the stdout
of the script. Handle the output however you need to (based on the
python script you wrote) and return it.